Auto-recover from "got state, but no session" errors
The easiest way to reproduce the problem is to try and `init()` with an
incorrect username/password.

Initialization would make some progress (and prepare some state files in
the DB directory), but would ultimately fail. With no session file
stored, these state files cannot be made use of later on.

We now have auto-recovery logic which detects this condition and purges
the database state files, so that a new session can be created cleanly.